requirements.txt
streamlit
python-dotenv
langchain  # if still needed
PyPDF2
chromadb
pdf2image
faiss-cpu
boto3  # Add the AWS SDK